add ability to run notebooks against another PAVICS host
This feature only works if the current hardcoded PAVICS host in the
notebooks is pavics.ouranos.ca.

This option has been chosen for the most seemless and transparent
experience for notebook authors.

The difficulty lies in the fact that all the outputs, which are static,
also need to be updated.  If we mandate the outputs have to be of the
form 'https://PAVICS_HOST/...' then the notebook author has to perform
extra contorsions to not display the real host for all cells, which also
affect readability.