launch container with current user UID and GID to not have write permission problems
Previously, those 2 launch scripts only worked on personal workstation where
the first created unix user (default 1000:1000) during installation is usually
the actually user.

But on a shared server, those 2 launch scripts will have write permissions
problems if the user running the script is not 1000:1000.

The fix is to change the jenkins user inside the container default UID and GID
of 1000:1000 to the UID and GID of the actual user running the script to
fix the write permissions problems.

No need to do anything under Jenkins since the docker image was specially made
to run with our Jenkins.  Under another Jenkins, a similar fix will need to be
applied.