allow to select pavics-sdi branch to test against
Needed to duplicate the default value for PAVICS_SDI_BRANCH in
default_build_params and testall and move it out off Jenkinsfile because
on the first Jenkins run, the build has no param and reference to
non-existing param in Jenkinsfile will break the build.

This is a well known limitation of declaring params inside a
Jenkinsfile.  The Jenkinsfile has to be read before the job knows it has
params but on the first build, the Jenkinsfile has not been read yet so
no params will exist.