save resulting notebooks to better help troubleshoot errors
The resulting notebooks is as if they were really run by the users will
end with .output.ipynb and will be saved as build artifacts by Jenkins.

This is another run so will double the run time and will not guaranty
exact same error as seen by py.test.

User can manually diff the original with the resulting notebooks this
way: nbdiff original.ipynb resulting.ipynb.output.ipynb (conda install
nbdime).

If this turns out to the a bad idea, we can simply change the default
value of the SAVE_RESULTING_NOTEBOOK boolean param in Jenkinsfile to
disable this behavior during automated build.

For manual run, runtest default to not perform this double run.

For manual Jenkins run, user can also disable this behavior.